Change table libraries; work on filtering/sorting
