+++ package.json
@@ -24,6 +24,7 @@
     "prop-types": "^15.7.2",
     "react": "^16.8.2",
     "react-bootstrap": "^1.0.0-beta.9",
+    "react-data-grid": "^6.1.0",
     "react-dom": "^16.8.2",
     "react-easy-swipe": "^0.0.18",
     "react-redux": "^7.1.0",
@@ -32,7 +33,6 @@
     "react-static-plugin-sass": "^7.1.0",
     "react-static-plugin-sitemap": "^7.0.0",
     "react-switch": "^5.0.0",
-    "react-table": "^6.10.0",
     "redux": "^4.0.1",
     "redux-logger": "^3.0.6",
     "redux-saga": "^1.0.4",
