+++ src/dynamic/AutoResponses/index.jsx
@@ -1,14 +1,23 @@
-import React, { useEffect } from "react";
+import React, { useEffect, useRef, useState, useMemo } from "react";
 import PropTypes from "prop-types";
 import { shallowEqual, useSelector } from "react-redux";
 import { getResponses } from "store/actions";
 import { useAuthDispatch, isNil } from "utility";
 
 import { Container } from "react-bootstrap";
-import ReactTable from "react-table";
+import ReactDataGrid from "react-data-grid";
 
 import "./style.scss";
-import "react-table/react-table.css";
+
+const baseColumnProps = {
+  sortable: true,
+  filterable: true
+};
+const columns = [
+  { key: "trigger", name: "Trigger", editable: true },
+  { key: "response", name: "Response", editable: true },
+  { key: "author_id", name: "Author" }
+].map(c => ({ ...c, ...baseColumnProps }));
 
 function AutoResponses({ guildId }) {
   // Connect to store
@@ -29,6 +38,66 @@ function AutoResponses({ guildId }) {
     if (authenticated) fetchResponses(guildId);
   }, [authenticated, guildId]);
 
+  // Escape hatch to access library methods imperatively
+  const dataGrid = useRef(null);
+
+  // Handle edits by propagating them to the store
+  const onGridRowsUpdated = ({ fromRow, toRow, updated }) => {
+    // console.log({ fromRow, toRow, updated });
+  };
+  // Open the editor upon cell selection
+  const [lastEditedPos, setLastEditedPos] = useState({ rowIdx: -1, idx: -1 });
+  const onCellSelected = ({ rowIdx, idx }) => {
+    dataGrid.current.openCellEditor(rowIdx, idx);
+    setLastEditedPos({ rowIdx, idx });
+  };
+  // Fix after-edit click
+  const onRowClick = (newRowIdx, _rowData, column) => {
+    if (isNil(column)) return;
+    const { idx, rowIdx } = lastEditedPos;
+    if (newRowIdx === rowIdx && column.idx === idx) {
+      dataGrid.current.openCellEditor(newRowIdx, column.idx);
+    }
+  };
+
+  // Row sorting
+  const [sortMeta, setSortMeta] = useState({
+    sortColumn: 0,
+    sortDirection: "NONE"
+  });
+  const onGridSort = (sortColumn, sortDirection) => {
+    setSortMeta({ sortColumn, sortDirection });
+  };
+  const sortRows = (rows, sortColumn, sortDirection) => {
+    const comparer = (a, b) => {
+      if (sortDirection === "ASC") {
+        return a[sortColumn] > b[sortColumn] ? 1 : -1;
+      } else if (sortDirection === "DESC") {
+        return a[sortColumn] < b[sortColumn] ? 1 : -1;
+      }
+    };
+    return sortDirection === "NONE" ? rows : [...rows].sort(comparer);
+  };
+  // Sorted view array
+  const rows = useMemo(
+    () => sortRows(commands, sortMeta.sortColumn, sortMeta.sortDirection),
+    [sortMeta, commands]
+  );
+
+  // Filtering
+  class AutoFilterToolbar extends React.Component {
+    componentDidMount() {
+      // eslint-disable-next-line react/prop-types
+      this.props.onToggleFilter();
+    }
+
+    render() {
+      // eslint-disable-next-line react/prop-types
+      return <div>{this.props.children}</div>;
+    }
+  }
+  // TODO make filtering functional
+
   return (
     <Container className="py-5 auto-responses">
       <h2>Automatic Responses</h2>
@@ -37,32 +106,18 @@ function AutoResponses({ guildId }) {
         current server.
       </p>
       <div className="table-outer">
-        <ReactTable
-          data={commands}
-          columns={[
-            {
-              Header: "Trigger",
-              accessor: "trigger"
-            },
-            {
-              Header: "Response",
-              accessor: "response"
-            },
-            {
-              Header: "Count",
-              accessor: "count"
-            },
-            {
-              Header: "Author",
-              id: "author",
-              accessor: d => {
-                const author = authors[d.author_id.toString()];
-                if (!isNil(author)) return author.name;
-                else return d.author_id;
-              }
-            }
-          ]}
-          defaultPageSize={100}
+        <ReactDataGrid
+          ref={dataGrid}
+          columns={columns}
+          rowGetter={i => rows[i]}
+          rowsCount={commands.length}
+          onGridRowsUpdated={onGridRowsUpdated}
+          onCellSelected={onCellSelected}
+          onRowClick={onRowClick}
+          enableCellSelect={true}
+          enableCellAutoFocus={false}
+          onGridSort={onGridSort}
+          toolbar={<AutoFilterToolbar />}
         />
       </div>
     </Container>
