+++ src/dynamic/Begin/style.scss
@@ -34,49 +34,42 @@
                 @include light {
                     opacity: 0.1;
                 }
-                @include theme-transitions("opacity");
             }
             .fg {
                 fill: #afafaf;
                 @include dark {
                     fill: color-adjust($-dark, 30%);
                 }
-                @include theme-transitions("fill");
             }
             .strong-fg {
                 fill: #4b4b4b;
                 @include dark {
                     fill: color-adjust($-dark, 50%);
                 }
-                @include theme-transitions("fill");
             }
             .strongest {
                 fill: #999999;
                 @include dark {
                     fill: color-adjust($-dark, -5%);
                 }
-                @include theme-transitions("fill");
             }
             .stronger {
                 fill: #c4c4c4;
                 @include dark {
                     fill: color-adjust($-dark, 5%);
                 }
-                @include theme-transitions("fill");
             }
             .strong {
                 fill: #efefef;
                 @include dark {
                     fill: color-adjust($-dark, 15%);
                 }
-                @include theme-transitions("fill");
             }
             .bg {
                 fill: #ffffff;
                 @include dark {
                     fill: color-adjust($-dark, 25%);
                 }
-                @include theme-transitions("fill");
             }
         }
     }
@@ -99,7 +92,6 @@
 
             @include dark-fg(rgba($-light, 0.6));
             @include light-fg(rgba($-dark, 0.6));
-            @include theme-transitions("color");
         }
 
         button {
