+++ src/utility/index.js
@@ -1,4 +1,9 @@
-export * from "utility/misc";
-export * from "utility/colors";
+export * from "utility/object";
+export * from "utility/array";
 export * from "utility/string";
+export * from "utility/colors";
 export * from "utility/hooks";
+export * from "utility/network";
+export * from "utility/document";
+
+export * from "utility/MockTyper";
